problem

+++ controllers/videoController.js
@@ -38,8 +38,11 @@ export const postUpload = async (req, res) => {
   const newVideo = await Video.create({
     fileUrl: path,
     title,
-    description
+    description,
+    creator: req.user.id
   });
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -48,7 +51,8 @@ export const videoDetail = async (req, res) => {
     params: { id }
   } = req;
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("creato");
+    console.log(video);
     res.render("videoDetail", { pagetitle: "videoDetail", video });
   } catch (error) {
     console.log(error);
@@ -62,6 +66,9 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
+    if (video.creator !== req.user.id) {
+      throw error;
+    }
     res.render("editVideo", { pagetitle: `Edit ${video.title}`, video });
   } catch (error) {
     res.redirect(routes.home);
@@ -86,6 +93,10 @@ export const deleteVideo = async (req, res) => {
     params: { id }
   } = req;
   try {
+    const video = await Video.findById(id);
+    if (video.creator !== req.user.id) {
+      throw error;
+    }
     await Video.findByIdAndRemove({ _id: id });
   } catch (error) {}
   res.redirect(routes.home);
